requirements.txt
flask
waitress
logging
requests
python-dotenv
open-webui